+++ ft_strlcpy.c
@@ -6,13 +6,13 @@
 /*   By: cschuijt <cschuijt@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/10/10 14:29:58 by cschuijt      #+#    #+#                 */
-/*   Updated: 2022/10/10 17:38:58 by cschuijt      ########   odam.nl         */
+/*   Updated: 2022/10/12 17:21:57 by cschuijt      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stddef.h>
 
-int	ft_strlen(const char *s);
+size_t	ft_strlen(const char *s);
 
 size_t	ft_strlcpy(char *dst, const char *src, size_t dstsize)
 {
