+++ ft_substr.c
@@ -6,14 +6,14 @@
 /*   By: cschuijt <cschuijt@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/10/04 13:54:42 by cschuijt      #+#    #+#                 */
-/*   Updated: 2022/10/08 21:30:07 by cschuijt      ########   odam.nl         */
+/*   Updated: 2022/10/12 17:22:05 by cschuijt      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stddef.h>
 #include <stdlib.h>
 
-int	ft_strlen(const char *s);
+size_t	ft_strlen(const char *s);
 
 char	*ft_substr(char const *s, unsigned int start, size_t len)
 {
